rename components for inverted relations
