+++ packages/client/src/pages/MainPage/containers/EntityDetailBox/EntityDetail/EntityDetailUsedInTable/EntityDetailInvertedRelations/EntityDetailInvertedRelation/EntityDetailInvertedRelationStyles.tsx
